Crate zng_tp_licenses

Third party license management and collection.

Crate

This crate is part of the zng project.

Cargo Features

This crate provides 2 feature flags, 0 enabled by default.

"build"

Include helpers for collecting third-party licenses.

"bundle"

Include helpers for deserializing build encoded licenses.

Macros

include_bundle

Includes the bundle file generated using write_bundle.

Structs

License

Represents a license id, name and text.

LicenseUsed

Represents a license and dependencies that use it.

User

Represents a project or package that uses a license.

UserLicense

Represents a license user with license.

Functions

collect_cargo_about

Calls cargo about for the crate.

decode_licenses

Decode licenses encoded with encode_licenses. Note that the encoded format is only guaranteed to work if both encoding and decoding is made with the same Cargo.lock dependencies.

encode_licenses

Bincode serialize and deflate the licenses.

merge_licenses

Merge licenses into into.

parse_cargo_about

Parse the output of cargo about.

sort_licenses

Sort vec by license name, and users of each license by name.

user_licenses

Invert data to be keyed by user, also sorts by user name.

write_bundle

Encode licenses and write to the output file that is included by include_bundle!.